+++ apps/xmtp.chat/src/components/App/App.tsx
@@ -1,33 +1,24 @@
 import { AppShell } from "@mantine/core";
 import { useDisclosure } from "@mantine/hooks";
 import { useEffect, useState } from "react";
-import { useLocation, useNavigate } from "react-router";
+import { useLocation } from "react-router";
 import { AppFooter } from "@/components/App/AppFooter";
 import { AppHeader } from "@/components/App/AppHeader";
 import { ErrorModal } from "@/components/App/ErrorModal";
 import { useAnalytics } from "@/hooks/useAnalytics";
-import { useClient } from "@/hooks/useClient";
+import { useRedirects } from "@/hooks/useRedirects";
 import { Main } from "@/routes/Main";
 import { Navbar } from "@/routes/Navbar";
 import classes from "./App.module.css";
 
 export const App: React.FC = () => {
+  useRedirects();
   useAnalytics();
   const [opened, { toggle }] = useDisclosure();
   const [collapsed, setCollapsed] = useState(true);
-  const { client } = useClient();
   const location = useLocation();
-  const navigate = useNavigate();
 
   useEffect(() => {
-    if (
-      !client &&
-      (location.pathname !== "/" || !location.pathname.startsWith("/dm"))
-    ) {
-      void navigate("/");
-      return;
-    }
-
     if (
       location.pathname.startsWith("/conversations") ||
       location.pathname.startsWith("/identity")
@@ -36,11 +27,7 @@ export const App: React.FC = () => {
     } else {
       setCollapsed(true);
     }
-
-    if (location.pathname === "/" && client) {
-      void navigate("/conversations");
-    }
-  }, [location.pathname, client, navigate]);
+  }, [location.pathname]);
 
   return (
     <>
@@ -60,7 +47,7 @@ export const App: React.FC = () => {
           <AppHeader opened={opened} toggle={toggle} collapsed={collapsed} />
         </AppShell.Header>
         <AppShell.Navbar className={classes.navbar}>
-          {client && <Navbar />}
+          <Navbar />
         </AppShell.Navbar>
         <AppShell.Main className={classes.main}>
           <Main />
